+++ tests/test_mechanics.py
@@ -517,7 +517,10 @@ def test_load():
         bounds = {"fix": fem.Boundary(field[0], fx=lambda x: x == 0)}
         dof0, dof1 = fem.dof.partition(field, bounds)
 
-        res = fem.newtonrhapson(items=[body, load], dof0=dof0, dof1=dof1)
+        def cb(dx, x, iteration, xnorm, fnorm, success):
+            print("fnorm", fnorm)
+
+        res = fem.newtonrhapson(items=[body, load], dof0=dof0, dof1=dof1, callback=cb)
 
         assert res.success