+++ tests/test_mesh.py
@@ -606,16 +606,16 @@ def test_expand_revolve_quadratic():
         .add_midpoints_edges()
         .add_midpoints_faces()
     )
-    
+
     new_mesh = mesh.revolve(n=3, phi=45)
     assert len(new_mesh.cells) == 8
-    
+
     new_mesh = mesh.revolve(n=11, phi=360)
     assert len(new_mesh.cells) == 4 * 10
-    
+
     new_mesh = mesh.revolve(phi=[0, 20, 40])
     assert len(new_mesh.cells) == 8
-    
+
     new_mesh = mesh.expand(z=[0, 1, 2])
     assert len(new_mesh.cells) == 8