+++ src/felupe/math/_math.py
@@ -1,4 +1,4 @@
-# -*- coding: utf-8 -*-
+s  # -*- coding: utf-8 -*-
 """
 This file is part of FElupe.