+++ CHANGELOG.md
@@ -20,6 +20,7 @@ All notable changes to this project will be documented in this file. The format 
 - Allow field containers to be included in the list of `fields` in `FieldContainer(fields)`. If a field container is included, its sub-fields are unpacked.
 - Set custom attributes by keyword-arguments in `FieldContainer(fields, **kwargs)`.
 - Show also the progress of Newton's method for `Job.evaluate(verbose=True)`.
+- Switch to `tqdm.auto` for better progress bars in Jupyter notebooks.
 
 ## [9.3.0] - 2025-06-29